build(package): refactoring build scripts with project ref / incremental (#5032)
* build(tsconfig): enable incremental, project references

* build(package): generate package using top level root as-is

* test(sideeffect): update import site snapshot

* build(package): support subpath alias with top-level root

* ci(circle): update ci scripts

* build(package): remove unused dependencies

* build(package): ensure alias script not throw

* test(mocha): run test directly from compiled results

* build(tsconfig): fix sourceroot for source, typemaps

BREAKING CHANGES: esm2015 renamed to esm. (though this is undocumented)
scoped package @reactivex/rxjs is being deprecated
