+++ .gitignore
@@ -19,10 +19,18 @@ img/
 spec-js/
 spec-build/
 .out/
+.tmp/
+
+ajax/
+fetch/
+operators/
+testing/
+webSocket/
 
 # Misc
 npm-debug.log
 .DS_STORE
+*.tgz
 
 # The check-side-effects package generates and deletes this file.
 # If the process is killed, it will be left behind.
