+++ integration/side-effects/side-effects.json
@@ -5,48 +5,48 @@
       "expectedOutput": "./snapshots/esm5/index.js"
     },
     {
-      "esModules": "../../dist/esm2015/index.js",
-      "expectedOutput": "./snapshots/esm2015/index.js"
+      "esModules": "../../dist/esm/index.js",
+      "expectedOutput": "./snapshots/esm/index.js"
     },
     {
       "esModules": "../../dist/esm5/ajax/index.js",
       "expectedOutput": "./snapshots/esm5/ajax.js"
     },
     {
-      "esModules": "../../dist/esm2015/ajax/index.js",
-      "expectedOutput": "./snapshots/esm2015/ajax.js"
+      "esModules": "../../dist/esm/ajax/index.js",
+      "expectedOutput": "./snapshots/esm/ajax.js"
     },
     {
       "esModules": "../../dist/esm5/fetch/index.js",
       "expectedOutput": "./snapshots/esm5/fetch.js"
     },
     {
-      "esModules": "../../dist/esm2015/fetch/index.js",
-      "expectedOutput": "./snapshots/esm2015/fetch.js"
+      "esModules": "../../dist/esm/fetch/index.js",
+      "expectedOutput": "./snapshots/esm/fetch.js"
     },
     {
       "esModules": "../../dist/esm5/operators/index.js",
       "expectedOutput": "./snapshots/esm5/operators.js"
     },
     {
-      "esModules": "../../dist/esm2015/operators/index.js",
-      "expectedOutput": "./snapshots/esm2015/operators.js"
+      "esModules": "../../dist/esm/operators/index.js",
+      "expectedOutput": "./snapshots/esm/operators.js"
     },
     {
       "esModules": "../../dist/esm5/testing/index.js",
       "expectedOutput": "./snapshots/esm5/testing.js"
     },
     {
-      "esModules": "../../dist/esm2015/testing/index.js",
-      "expectedOutput": "./snapshots/esm2015/testing.js"
+      "esModules": "../../dist/esm/testing/index.js",
+      "expectedOutput": "./snapshots/esm/testing.js"
     },
     {
       "esModules": "../../dist/esm5/webSocket/index.js",
       "expectedOutput": "./snapshots/esm5/websocket.js"
     },
     {
-      "esModules": "../../dist/esm2015/webSocket/index.js",
-      "expectedOutput": "./snapshots/esm2015/websocket.js"
+      "esModules": "../../dist/esm/webSocket/index.js",
+      "expectedOutput": "./snapshots/esm/websocket.js"
     }
   ]
 }
