+++ integration/side-effects/snapshots/esm/fetch.js
@@ -1,0 +1,1 @@
+
