+++ integration/side-effects/snapshots/esm/index.js
@@ -1,5 +1,3 @@
-import "tslib";
-
 var NotificationKind;
 
 (function(NotificationKind) {
