+++ integration/side-effects/snapshots/esm/operators.js
@@ -1,5 +1,3 @@
-import "tslib";
-
 var NotificationKind;
 
 (function(NotificationKind) {
