+++ integration/side-effects/snapshots/esm/testing.js
@@ -1,5 +1,3 @@
-import "tslib";
-
 var NotificationKind;
 
 (function(NotificationKind) {
