+++ integration/side-effects/snapshots/esm5/websocket.js
@@ -1,1 +1,9 @@
 import "tslib";
+
+var NotificationKind;
+
+(function(NotificationKind) {
+    NotificationKind["NEXT"] = "N";
+    NotificationKind["ERROR"] = "E";
+    NotificationKind["COMPLETE"] = "C";
+})(NotificationKind || (NotificationKind = {}));
