+++ spec/tsconfig.json
@@ -1,3 +1,14 @@
 {
-  "extends": "../tsconfig.json"
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "noEmit": false,
+    "outDir": "../dist/spec",
+    "target": "es5",
+    "module": "commonjs"
+  },
+  "references": [
+    {
+      "path": "../src/tsconfig.types.json"
+    }
+  ]
 }
