--- src/fetch/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "rxjs/fetch",
-  "typings": "./index.d.ts",
-  "main": "./index.js",
-  "module": "../_esm5/fetch/index.js",
-  "es2015": "../_esm2015/fetch/index.js",
-  "sideEffects": false
-}
