--- src/operators/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "rxjs/operators",
-  "typings": "./index.d.ts",
-  "main": "./index.js",
-  "module": "../_esm5/operators/index.js",
-  "es2015": "../_esm2015/operators/index.js",
-  "sideEffects": false
-}
