+++ src/tsconfig.cjs.json
@@ -4,5 +4,8 @@
     "module": "commonjs",
     "target": "es5",
     "outDir": "../dist/cjs"
-  }
+  },
+  "exclude": [
+    "./internal/umd.ts"
+  ]
 }
