+++ src/tsconfig.esm.json
@@ -1,10 +1,9 @@
 {
   "extends": "./tsconfig.base.json",
   "compilerOptions": {
-    "module": "es2015",
+    "module": "esnext",
     "importHelpers": true,
-    "moduleResolution": "node",
     "target": "es2015",
-    "outDir": "../dist/esm2015"
+    "outDir": "../dist/esm"
   }
 }
