+++ src/tsconfig.esm5.rollup.json
@@ -3,5 +3,6 @@
   "compilerOptions": {
     "outDir": "../dist/esm5_for_rollup",
     "noEmitHelpers": true
-  }
+  },
+  "exclude": []
 }
