+++ src/tsconfig.types.json
@@ -1,0 +1,16 @@
+{
+  "extends": "./tsconfig.esm.json",
+  "compilerOptions": {
+    "composite": true,
+    "module": "es2015",
+    "target": "esnext",
+    "removeComments": false,
+    "declaration": true,
+    "declarationMap": true,
+    "declarationDir": "../dist/types",
+    "emitDeclarationOnly": true
+  },
+  "exclude": [
+    "./internal/umd.ts"
+  ]
+}
