--- src/webSocket/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "rxjs/webSocket",
-  "typings": "./index.d.ts",
-  "main": "./index.js",
-  "module": "../_esm5/webSocket/index.js",
-  "es2015": "../_esm2015/webSocket/index.js",
-  "sideEffects": false
-}
