--- tsconfig.base.json
@@ -1,28 +1,0 @@
-{
-  "compilerOptions": {
-    "removeComments": true,
-    "preserveConstEnums": true,
-    "sourceMap": true,
-    "strictFunctionTypes": true,
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "suppressImplicitAnyIndexErrors": true,
-    "moduleResolution": "node",
-    "stripInternal": false,
-    "target": "es5",
-    "outDir": "./.out",
-    "lib": [
-      "es5",
-      "es2018",
-      "dom"
-    ]
-  },
-  "formatCodeOptions": {
-    "indentSize": 2,
-    "tabSize": 2
-  },
-  "bazelOptions": {
-    "suppressTsconfigOverrideWarnings": true
-  }
-}
