--- tsconfig/tsconfig.base.json
@@ -1,24 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "baseUrl": "../src",
-    "sourceRoot": "../src",
-    "paths": {
-      "rxjs": ["./"],
-      "rxjs/*": ["./*"]
-    }
-  },
-  "files": [
-    // entry-points
-    "../src/index.ts",
-    "../src/ajax/index.ts",
-    "../src/fetch/index.ts",
-    "../src/operators/index.ts",
-    "../src/testing/index.ts",
-    "../src/webSocket/index.ts",
-
-    // umd entry-point
-    "../src/internal/umd.ts"
-  ]
-}
